use _error.ts to avoid `node:` side effects
